moved fetch call to homepage
